Made checkboxes more accurate!
Designed a new vector for them, to replace the old, janky Unicode character setup with a new, cleaner ::before and ::after setup that only requires one input.
In the process, also removed the stupid requirement for a "webtv-checkbox" div.
Also attempted to fix a bug where the status bar doesn't appear on some browsers.