Added search page! + Fixed splash logo
Changes:
Split up Template.html into 2: One for pages that remain in the root, and one for service pages that go in a folder.
Changed the sidebar text template to be easier to edit.
Added stuff to the audioscope testing page.
Fixed splash logo centering.
Updated the dialing theme to use webcd's sweet HD remaster.

Co-Authored-By: nitrate92 <46730268+nitrate92@users.noreply.github.com>